# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A lightweight, read-only Kubernetes service that provides a RESTful API for querying pod and node information\n\nThe K8s Cluster Agent runs inside a Kubernetes cluster and provides various endpoints for:\n- Pod information (description, scheduling, resources, health)\n- Node utilization metrics\n- Cluster-wide pod issues dashboard\n- Namespace error analysis",
        "title": "K8s Cluster Agent API",
        "contact": {
            "name": "K8s Cluster Agent Team",
            "url": "https://github.com/sumandas0/k8s-cluster-agent"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/pods/{namespace}/{podName}/describe": {
            "get": {
                "description": "Returns comprehensive pod information including status, containers, volumes, and conditions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pods"
                ],
                "summary": "Get pod description",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace name",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pod name",
                        "name": "podName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pod description",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pod not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Request timeout",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pods/{namespace}/{podName}/failure-events": {
            "get": {
                "description": "Returns categorized and analyzed failure events with insights and recommendations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pods"
                ],
                "summary": "Get pod failure events analysis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace name",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pod name",
                        "name": "podName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pod failure events analysis",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.PodFailureEvents"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pod not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Request timeout",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/pods/{namespace}/{podName}/resources": {
            "get": {
                "description": "Returns detailed resource requirements (CPU, memory) for all containers in the pod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pods"
                ],
                "summary": "Get pod resource information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace name",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pod name",
                        "name": "podName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pod resource information",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.PodResources"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pod not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Request timeout",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pods/{namespace}/{podName}/scheduling": {
            "get": {
                "description": "Returns comprehensive scheduling analysis including placement decisions, failure reasons, and constraints",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pods"
                ],
                "summary": "Get pod scheduling information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace name",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pod name",
                        "name": "podName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pod scheduling information",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.PodScheduling"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pod not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Request timeout",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pods/{namespace}/{podName}/scheduling/explain": {
            "get": {
                "description": "Returns Elasticsearch-style detailed explanations for pod scheduling decisions with per-node analysis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pods"
                ],
                "summary": "Get detailed pod scheduling explanation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace name",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pod name",
                        "name": "podName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pod scheduling explanation",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.SchedulingExplanation"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pod not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Request timeout",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.PodFailureEvents": {
            "type": "object",
            "properties": {
                "criticalEvents": {
                    "type": "integer"
                },
                "eventCategories": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "failureEvents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.FailureEvent"
                    }
                },
                "mostRecentIssue": {
                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.FailureEvent"
                },
                "namespace": {
                    "type": "string"
                },
                "ongoingIssues": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "podName": {
                    "type": "string"
                },
                "podPhase": {
                    "type": "string"
                },
                "podStatus": {
                    "type": "string"
                },
                "totalEvents": {
                    "type": "integer"
                },
                "warningEvents": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.PodResources": {
            "type": "object",
            "properties": {
                "containers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.ContainerResources"
                    }
                },
                "total": {
                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.ResourceSummary"
                }
            }
        },
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.PodScheduling": {
            "type": "object",
            "properties": {
                "affinity": {
                    "$ref": "#/definitions/v1.Affinity"
                },
                "conditions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.PodCondition"
                    }
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.SchedulingEvent"
                    }
                },
                "failureCategories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.SchedulingFailureCategory"
                    }
                },
                "failureSummary": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.FailureCategorySummary"
                    }
                },
                "nodeName": {
                    "type": "string"
                },
                "nodeSelector": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "priority": {
                    "type": "integer"
                },
                "priorityClassName": {
                    "type": "string"
                },
                "schedulerName": {
                    "type": "string"
                },
                "schedulingDecisions": {
                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.SchedulingDecisions"
                },
                "status": {
                    "type": "string"
                },
                "tolerations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.Toleration"
                    }
                },
                "unschedulableNodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.UnschedulableNode"
                    }
                }
            }
        },
...
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.SchedulingExplanation": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.SchedulingEvent"
                    }
                },
                "namespace": {
                    "type": "string"
                },
                "nodeAnalysis": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.NodeSchedulingExplanation"
                    }
                },
                "nodeName": {
                    "type": "string"
                },
                "podName": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "summary": {
                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.SchedulingSummary"
                }
            }
        },
...
        "github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorDetail"
                },
                "metadata": {
                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.Metadata"
                }
            }
        },
        "github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.Metadata": {
            "type": "object",
            "properties": {
                "requestId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "metadata": {
                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.Metadata"
                }
            }
        },
...
    },
...
}